queryVisibleFeatures
